% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceiling2.R
\name{ceiling2}
\alias{ceiling2}
\alias{floor2}
\title{rounding of numbers}
\usage{
ceiling2(x, digits = 1)

floor2(x, digits = 1)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{integer indicating the number of significant digits.}
}
\value{
ceiling2 rounds the elements in x to the specified number of significant digits that is the smallest number not less than the corresponding elements.

floor2 rounds the elements in x to the specified number of significant digits that is the largest number not greater than the corresponding elements.
}
\description{
The ceiling2 is ceiling of numeric values by digits. The floor2 is floor of numeric values by digits.
}
\examples{
x = c(12345, 54.321)

ceiling2(x)
ceiling2(x, 2)
ceiling2(x, 3)

floor2(x)
floor2(x, 2)
floor2(x, 3)

}
