% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cols_const}
\alias{cols_const}
\title{constant columns}
\usage{
cols_const(dt)
}
\arguments{
\item{dt}{a data frame.}
}
\description{
The columns name of a data frame with constant value.
}
\examples{
dt = data.frame(a = sample(0:9, 6), b = sample(letters, 6),
                c = rep(1, 6), d = rep('a', 6))
dt
cols_const(dt)

}
