% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funcs.R
\name{date_bop}
\alias{date_bop}
\alias{date_eop}
\title{start/end date by period}
\usage{
date_bop(x, freq, workday = FALSE)

date_eop(x, freq, workday = FALSE)
}
\arguments{
\item{x}{a date}

\item{freq}{the frequency of period. It supports weekly, monthly, quarterly and yearly.}

\item{workday}{logical, whether to return the latest workday}
}
\description{
The date of bop (beginning of period) or eop (end of period).
}
