% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funcs.R
\name{date_from}
\alias{date_from}
\title{start date by range}
\usage{
date_from(date_range, to = Sys.Date(), default_from = "1000-01-01")
}
\arguments{
\item{date_range}{date range, available value including nd, nm, mtd, qtd, ytd, ny, max.}

\item{to}{a date, default is current system date.}

\item{default_from}{the default date when date_range is sett to max}
}
\value{
It returns the start date of a date_range with a specified end date.
}
\description{
The date before a specified date by date_range.
}
\examples{
date_from(3)
date_from('3d')

date_from('3m')
date_from('3q')
date_from('3y')

date_from('mtd')
date_from('qtd')
date_from('ytd')

}
