% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funcs.R
\name{date_num}
\alias{date_num}
\title{date to number}
\usage{
date_num(x, unit = "s", origin = "1970-01-01")
}
\arguments{
\item{x}{date.}

\item{unit}{number unit, available values including ms (milliseconds), s (seconds), d (days).}

\item{origin}{original date, defaults to 1970-01-01.}
}
\description{
It converts date to a number of milliseconds/seconds/days since an original date.
}
\examples{
date_num(Sys.time(), unit='ms')
date_num(Sys.time(), unit='s')
date_num(Sys.time(), unit='d')

}
