% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max2.R
\name{max2}
\alias{max2}
\alias{min2}
\title{Maxima and Minima}
\usage{
max2(..., na.rm = FALSE)

min2(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric or character arguments}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\description{
Returns the (regular or parallel) maxima and minima of the input values. For numeric NAs, it returns NA instead of Inf or -Inf.
}
\examples{
max2(c(NA), na.rm=TRUE)
max(c(NA), na.rm=TRUE)

min2(c(NA), na.rm=TRUE)
min(c(NA), na.rm=TRUE)

}
