% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge2}
\alias{merge2}
\title{merge data.frames list}
\usage{
merge2(datlst, by = NULL, all = TRUE, ...)
}
\arguments{
\item{datlst}{a list of data.frames.}

\item{by}{A vector of shared column names in x and y to merge on. This defaults to the shared key columns between the two tables. If y has no key columns, this defaults to the key of x.}

\item{all}{logical; all = TRUE is shorthand to save setting both all.x = TRUE and all.y = TRUE.}

\item{...}{Additional parameters provided in the merge function.}
}
\description{
Merge a list of data.frames by common columns or row names.
}
