% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprate.R
\name{reprate}
\alias{reprate}
\title{char repetition rate}
\usage{
reprate(dt, col)
}
\arguments{
\item{dt}{a data frame or vector.}

\item{col}{a character column name.}
}
\description{
reprate estimate the max rate of character repetition.
}
\examples{
x = c('a', 'aa', 'ab', 'aab', 'aaab')
reprate(x)

reprate(data.frame(x=x), 'x')

}
