% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split2}
\alias{split2}
\title{split vector by equal size}
\usage{
split2(x, n)
}
\arguments{
\item{x}{a vector.}

\item{n}{a numeric, size of n.}
}
\description{
Split vector x into chunks of equal size n
}
\examples{
x = 1:9

split2(x, 3)
split2(x, 6)

}
