% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conticnt.R
\name{conticnt}
\alias{conticnt}
\title{continuous counting}
\usage{
conticnt(dt, col, opr = "==", threshold = 0, cnt = FALSE)
}
\arguments{
\item{dt}{a data frame or vector.}

\item{col}{a numeric column name.}

\item{opr}{relational operators, defaults to ==.}

\item{threshold}{threshold value, defaults to 0.}

\item{cnt}{whether to count the number rows in each continuous groups.}
}
\description{
It counts the number of values continuous comparing with a threshold.
}
\examples{
# example I
dt1 = c(0,0,0, 1,1,1, 0,0, 1,1)
conticnt(dt1)
conticnt(dt1, cnt = TRUE)
conticnt(dt1, opr = '<=')

# example II
dt2 = data.frame(x = c(0,0,0, 1,1,1, 0,0, 1,1))
conticnt(dt2, col='x')
conticnt(dt2, col='x', cnt=TRUE)
conticnt(dt2, col='x', opr='<=')

library(data.table)
setDT(dt2)[, ct_N := conticnt(x)][]

# example III
# number of growth values
dt3 = data.frame(x = c(0,1,2, 3,2,1, 0,1, 2,1))
dt3 = setDT(dt3)[, diff := x - shift(x,type='lag',fill = 0)]
conticnt(dt3, col='diff', opr = '>=')

}
