% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funcs.R
\name{date_bop}
\alias{date_bop}
\alias{date_eop}
\title{start/end date by period}
\usage{
date_bop(freq, x, workday = FALSE)

date_eop(freq, x, workday = FALSE)
}
\arguments{
\item{freq}{the frequency of period. It supports weekly, monthly, quarterly and yearly.}

\item{x}{a date}

\item{workday}{logical, whether to return the latest workday}
}
\value{
date_bop returns the beginning date of period of corresponding x by frequency.

date_eop returns the end date of period of corresponding x by frequency.
}
\description{
The date of bop (beginning of period) or eop (end of period).
}
\examples{
date_bop('weekly', Sys.Date())
date_eop('weekly', Sys.Date())

date_bop('monthly', Sys.Date())
date_eop('monthly', Sys.Date())

}
