% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprate.R
\name{reprate}
\alias{reprate}
\title{char repetition rate}
\usage{
reprate(x, col)
}
\arguments{
\item{x}{a character vector or a data frame.}

\item{col}{a character column name.}
}
\value{
a numeric vector indicating the max rate of character repetition in the corresponding elements in argument x vector.
}
\description{
reprate estimates the max rate of character repetition.
}
\examples{
x = c('a', 'aa', 'ab', 'aab', 'aaab')
reprate(x)

reprate(data.frame(x=x), 'x')

}
